requirements.txt
Django==1.11.4
Pillow==4.2.1
argparse==1.2.1
boto==2.48.0
boto3==1.4.7
botocore==1.7.20
django-ckeditor==5.3.0
django-cors-headers==1.1.0
django-extensions==1.7.4
django-filter==1.0.4
django-grappelli==2.10.1
django-js-asset==0.1.1
django-model-utils==2.6
django-storages==1.6.5
docutils==0.14
futures==3.1.1
gunicorn==19.3.0
jmespath==0.9.3
olefile==0.44
psycopg2==2.7.3
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
s3transfer==0.1.11
six==1.10.0
sorl-thumbnail==12.3
whitenoise==3.3.0
wsgiref==0.1.2
